add caching of secure-file

+++ scripts/ps/appveyor_install.ps1
@@ -12,8 +12,13 @@ function Check-ExitCode
 
 if ($env:ENCKEY)
 {
-    Write-Host "Downloading secure-file from NuGet" -ForegroundColor Cyan
-    nuget install secure-file -ExcludeVersion
+    if (-not (Test-Path "secure-file"))
+    {
+        Write-Host "Downloading secure-file from NuGet..." -ForegroundColor Cyan
+        nuget install secure-file -ExcludeVersion
+    }
+    
+    Write-Host "Decrypting code signing key..."            -ForegroundColor Cyan
     secure-file\tools\secure-file -decrypt distribution\windows\code-sign-key-openrct2.org.pfx.enc -secret $env:ENCKEY
     Check-ExitCode
 }
@@ -22,7 +27,7 @@ if (${env:OPENRCT2.ORG_TOKEN})
 {
     if (-not (Test-Path "C:\ProgramData\chocolatey\lib\nsis.portable"))
     {
-        Write-Host "Downloading NSIS from chocolatey"   -ForegroundColor Cyan
+        Write-Host "Downloading NSIS from chocolatey..."   -ForegroundColor Cyan
         cinst nsis.portable --version=3.01-beta1 > $null
         if ($LASTEXITCODE -ne 0)
         {
@@ -30,7 +35,7 @@ if (${env:OPENRCT2.ORG_TOKEN})
             $host.SetShouldExit(1)
         }
 
-        Write-Host "Downloading KillProcDLL for NSIS"   -ForegroundColor Cyan
+        Write-Host "Downloading KillProcDLL for NSIS..."   -ForegroundColor Cyan
         curl "http://nsis.sourceforge.net/mediawiki/images/5/53/KillProcDll%26FindProcDll.zip" -OutFile nsisxtra.zip
         Check-ExitCode
         7z x nsisxtra.zip > $null
